+++ src/pages/Profile/EditProfilePage.tsx
@@ -12,6 +12,7 @@ import { transformToEditProfileData } from '@/utils/editProfileData';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { country, phone, visa } from '@/constants/information';
+import useNavigateBack from '@/hooks/useNavigateBack';
 
 const EditProfilePage = () => {
   const navigate = useNavigate();
@@ -38,13 +39,7 @@ const EditProfilePage = () => {
       });
     };
 
-  const handleBackButtonClick = () => {
-    if (window.history.length > 1) {
-      navigate(-1);
-    } else {
-      navigate('/');
-    }
-  };
+  const handleBackButtonClick = useNavigateBack();
 
   const handleSubmit = async (): Promise<void> => {
     // TODO : API - 3.5 (유학생) 프로필 수정
