--- src/pages/SetIntroduction/IntroductionPage.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const IntroductionPage = () => {
-  return <div>IntroductionPage</div>;
-};
-
-export default IntroductionPage;
